- Fixed: problems with shell extensions [added extra AddRef (which may perhaps cause problems, let see)] - Fixed: date format dd-MMM-yyyy didn't work - Fixed: in locate.ini, FILE is assumed to be in the same directory as locate.ini if path not given

+++ Locate/Locate32/ResultsDialogs.cpp
@@ -1269,7 +1269,7 @@ BOOL CSaveResultsDlg::OnFileNameOK()
 	if (IsDlgButtonChecked(IDC_SELECTEDITEMS))
 		m_nFlags|=RESULT_INCLUDESELECTEDITEMS;
 
-	int nTemplateSel=SendDlgItemMessage(IDC_TEMPLATE,CB_GETCURSEL)-1;
+	int nTemplateSel=(int)SendDlgItemMessage(IDC_TEMPLATE,CB_GETCURSEL)-1;
 	if (nTemplateSel>=0 && nTemplateSel<m_TemplateFiles.GetSize())
 		m_strTemplate=m_TemplateFiles[nTemplateSel];
 	else
